% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possession_chains.R
\name{add_next_chain_outcome}
\alias{add_next_chain_outcome}
\title{Add Opponent's Next Chain Outcome}
\usage{
add_next_chain_outcome(chain_outcomes)
}
\arguments{
\item{chain_outcomes}{Data frame from classify_chain_outcomes()}
}
\value{
Chain outcomes with next_chain_goal column added
}
\description{
For EPV conceding model, we need to know if the opponent scores on their
next possession after we lose the ball. Optimized with data.table.
}
