% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_features.R
\name{add_possession_velocity}
\alias{add_possession_velocity}
\title{Calculate Possession Velocity}
\usage{
add_possession_velocity(spadl_with_chains)
}
\arguments{
\item{spadl_with_chains}{SPADL actions with chain information}
}
\value{
SPADL actions with velocity features added
}
\description{
Computes the speed of ball progression in a possession chain.
Optimized with data.table.
}
