% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{add_xpass_to_spadl}
\alias{add_xpass_to_spadl}
\title{Add xPass to SPADL Actions}
\usage{
add_xpass_to_spadl(spadl_actions, xpass_model)
}
\arguments{
\item{spadl_actions}{SPADL actions data frame}

\item{xpass_model}{Fitted xPass model}
}
\value{
SPADL actions with xpass column added for passes
}
\description{
Adds pass completion probability to pass actions in SPADL data.
}
\keyword{internal}
