% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_model.R
\name{aggregate_rapm_by_team}
\alias{aggregate_rapm_by_team}
\title{Aggregate RAPM by team}
\usage{
aggregate_rapm_by_team(ratings, player_data)
}
\arguments{
\item{ratings}{Data frame of player ratings}

\item{player_data}{Data frame linking players to teams}
}
\value{
Data frame with team-level summaries
}
\description{
Summarizes player ratings at team level.
}
\keyword{internal}
