% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{assign_players_to_splints_fast}
\alias{assign_players_to_splints_fast}
\title{Assign players to splints (fast version)}
\usage{
assign_players_to_splints_fast(boundaries, lineups, match_id)
}
\description{
Assign players to splints (fast version)
}
\keyword{internal}
