% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{calculate_action_type_epv}
\alias{calculate_action_type_epv}
\title{Calculate EPV by Action Type}
\usage{
calculate_action_type_epv(spadl_with_epv)
}
\arguments{
\item{spadl_with_epv}{SPADL actions with EPV}
}
\value{
Data frame with EPV totals by action type per player
}
\description{
Calculate EPV by Action Type
}
\keyword{internal}
