% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{calculate_defensive_spm}
\alias{calculate_defensive_spm}
\title{Calculate defensive SPM}
\usage{
calculate_defensive_spm(data, defensive_cols = NULL, alpha = 0.5)
}
\arguments{
\item{data}{SPM regression data}

\item{defensive_cols}{Defensive predictor columns}

\item{alpha}{Elastic net mixing}
}
\value{
Fitted model for defensive SPM
}
\description{
Fits SPM model for defensive contribution only.
}
\keyword{internal}
