% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{calculate_distance_to_goal}
\alias{calculate_distance_to_goal}
\title{Calculate Distance to Goal}
\usage{
calculate_distance_to_goal(x, y, goal_x = 100, goal_y = 50)
}
\arguments{
\item{x}{X coordinate (0-100 scale)}

\item{y}{Y coordinate (0-100 scale)}

\item{goal_x}{X coordinate of goal center (default 100)}

\item{goal_y}{Y coordinate of goal center (default 50)}
}
\value{
Distance to goal in coordinate units
}
\description{
Calculates Euclidean distance from coordinates to center of goal.
Assumes goal is at x=100, y=50.
}
\examples{
calculate_distance_to_goal(85, 50)  # Close, central
calculate_distance_to_goal(50, 50)  # Halfway line
}
