% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{calculate_effective_minute}
\alias{calculate_effective_minute}
\title{Calculate effective minute including stoppage time}
\usage{
calculate_effective_minute(minute, added_time, first_half_stoppage = 0L)
}
\arguments{
\item{minute}{Integer vector of base minutes}

\item{added_time}{Integer vector of added time (can be NA)}

\item{first_half_stoppage}{Integer, minutes of first-half stoppage (default 0)}
}
\value{
Numeric vector of effective minutes
}
\description{
Converts minute + added_time into a single continuous time value.
Second-half events are offset by first-half stoppage to create continuous time.
}
\details{
Examples with 3 mins first-half stoppage:
\itemize{
\item First half minute 30 -> 30
\item First half 45+3 -> 48
\item Second half minute 46 -> 49 (offset by 3)
\item Second half 90+11 -> 104 (90 + 11 + 3)
}
}
\keyword{internal}
