% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calculate_npxg}
\alias{calculate_npxg}
\title{Calculate non-penalty xG and goals}
\usage{
calculate_npxg(shooting)
}
\arguments{
\item{shooting}{Processed shooting data}
}
\value{
Data frame with npxG by player and match
}
\description{
Calculates non-penalty expected goals from shot data.
}
\keyword{internal}
