% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{calculate_offensive_spm}
\alias{calculate_offensive_spm}
\title{Calculate offensive SPM}
\usage{
calculate_offensive_spm(data, offensive_cols = NULL, alpha = 0.5)
}
\arguments{
\item{data}{SPM regression data}

\item{offensive_cols}{Offensive predictor columns}

\item{alpha}{Elastic net mixing}
}
\value{
Fitted model for offensive SPM
}
\description{
Fits SPM model for offensive contribution only.
}
\keyword{internal}
