% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{calculate_splint_npxgd}
\alias{calculate_splint_npxgd}
\title{Calculate splint-level xG differential}
\usage{
calculate_splint_npxgd(
  boundaries,
  shooting,
  match_id,
  home_team = NULL,
  away_team = NULL
)
}
\arguments{
\item{boundaries}{Data frame of splint boundaries}

\item{shooting}{Processed shooting data}

\item{match_id}{Match identifier}

\item{home_team}{Name of the home team}

\item{away_team}{Name of the away team}
}
\value{
Data frame with npxGD for each splint
}
\description{
Calculates the non-penalty xG differential for each splint.
}
\keyword{internal}
