% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{calculate_splint_npxgd_fast}
\alias{calculate_splint_npxgd_fast}
\title{Calculate splint xG (fast version)}
\usage{
calculate_splint_npxgd_fast(boundaries, shooting, home_team, away_team)
}
\description{
Calculate splint xG (fast version)
}
\keyword{internal}
