% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{calculate_spm_ratings_xgb}
\alias{calculate_spm_ratings_xgb}
\title{Calculate SPM ratings using XGBoost model}
\usage{
calculate_spm_ratings_xgb(player_features, spm_xgb_model)
}
\arguments{
\item{player_features}{Data frame of player features}

\item{spm_xgb_model}{Fitted XGBoost SPM model from fit_spm_xgb}
}
\value{
Data frame with SPM ratings
}
\description{
Calculate SPM ratings using XGBoost model
}
\seealso{
Other spm: 
\code{\link{calculate_spm_ratings}()},
\code{\link{fit_spm_model}()},
\code{\link{get_spm_feature_importance}()}
}
\concept{spm}
