% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAIN_FINAL_THIRD_X}
\alias{CHAIN_FINAL_THIRD_X}
\title{Final third x-coordinate threshold}
\format{
Numeric value: 66
}
\usage{
CHAIN_FINAL_THIRD_X
}
\description{
Actions with start_x greater than this value are considered
to be in the attacking final third (0-100 pitch scale).
}
\keyword{internal}
