% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAIN_PROGRESSIVE_THRESHOLD}
\alias{CHAIN_PROGRESSIVE_THRESHOLD}
\title{Progressive chain distance threshold}
\format{
Numeric value: 25
}
\usage{
CHAIN_PROGRESSIVE_THRESHOLD
}
\description{
Minimum forward x-distance (in 0-100 units) a player must move
the ball within a chain to count as a "progressive" chain contribution.
}
\keyword{internal}
