% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possession_chains.R
\name{classify_chain_outcomes}
\alias{classify_chain_outcomes}
\title{Classify Chain Outcomes}
\usage{
classify_chain_outcomes(spadl_with_chains)
}
\arguments{
\item{spadl_with_chains}{SPADL actions with chain_id from create_possession_chains()}
}
\value{
Data frame of chain-level statistics
}
\description{
Determines the outcome of each possession chain (goal, shot, turnover, etc.)
Optimized with data.table.
}
