% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_prediction.R
\name{compute_multiclass_logloss}
\alias{compute_multiclass_logloss}
\title{Compute Multi-Class Log Loss}
\usage{
compute_multiclass_logloss(y_true, prob_matrix, eps = 1e-15)
}
\arguments{
\item{y_true}{Integer vector of true labels (0, 1, 2)}

\item{prob_matrix}{Matrix with 3 columns (P(0), P(1), P(2))}

\item{eps}{Clipping epsilon to avoid log(0) (default 1e-15)}
}
\value{
Scalar log loss value
}
\description{
Compute Multi-Class Log Loss
}
\seealso{
Other match prediction: 
\code{\link{aggregate_lineup_ratings}()},
\code{\link{aggregate_lineup_skills}()},
\code{\link{augment_ratings_with_history}()},
\code{\link{calibration_table}()},
\code{\link{compute_match_elos}()},
\code{\link{compute_team_rolling_features}()},
\code{\link{init_team_elos}()},
\code{\link{predict_match}()},
\code{\link{update_elo}()}
}
\concept{match prediction}
