% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_features.R
\name{create_location_features}
\alias{create_location_features}
\title{Create Location-Only Features}
\usage{
create_location_features(x, y)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}
}
\value{
Data frame with location features
}
\description{
Creates a minimal feature set based only on location.
}
