% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{create_match_splints}
\alias{create_match_splints}
\title{Create splints for a single match}
\usage{
create_match_splints(
  match_id,
  events,
  lineups,
  shooting,
  results,
  include_goals = TRUE
)
}
\arguments{
\item{match_id}{Match identifier}

\item{events}{Processed events data}

\item{lineups}{Processed lineups data}

\item{shooting}{Processed shooting data}

\item{results}{Processed match results}

\item{include_goals}{Whether to create splints at goal times}
}
\value{
List with splint data for the match
}
\description{
Master function to create all splint data for one match.
}
