% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possession_chains.R
\name{create_possession_chains}
\alias{create_possession_chains}
\title{Create Possession Chains from SPADL Actions}
\usage{
create_possession_chains(spadl_actions)
}
\arguments{
\item{spadl_actions}{Data frame in SPADL format from convert_opta_to_spadl()}
}
\value{
Data frame with chain assignments added
}
\description{
Groups consecutive actions by the same team into possession chains.
Optimized with data.table for fast processing of large datasets.
}
\details{
A chain ends when:
\itemize{
\item Opponent gains possession (successful tackle/interception)
\item Ball goes out of play (foul, throw-in situation)
\item Goal is scored
\item Period ends
\item Large time gap (>30 seconds)
}
}
\examples{
\dontrun{
spadl <- convert_opta_to_spadl(opta_events)
chains <- create_possession_chains(spadl)
head(chains[, c("match_id", "chain_id", "action_type", "team_id")])
}
}
\seealso{
Other possession chains: 
\code{\link{add_red_card_to_chains}()}
}
\concept{possession chains}
