% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{create_splint_boundaries}
\alias{create_splint_boundaries}
\title{Create splint boundaries}
\usage{
create_splint_boundaries(events, include_goals = TRUE, include_halftime = TRUE)
}
\arguments{
\item{events}{Data frame of match events}

\item{include_goals}{Logical, whether to create new splints at goals (default TRUE)}

\item{include_halftime}{Logical, whether to create splint at halftime (default TRUE)}
}
\value{
Data frame of splint boundaries with game state
}
\description{
Defines start and end times for each splint based on events.
Splints are created at:
\itemize{
\item Start of match (minute 0)
\item Each goal
\item Each substitution
\item Each red card
\item Half time (minute 45)
\item End of match (minute 90)
}
}
\details{
Also tracks game state (cumulative goals, red cards, player counts) at each splint boundary.
}
