% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_utils.R
\name{derive_events_from_shooting}
\alias{derive_events_from_shooting}
\title{Derive goal events from shooting data}
\usage{
derive_events_from_shooting(shooting_data)
}
\arguments{
\item{shooting_data}{Shooting data from load_fb_match_shooting (already snake_case)}
}
\value{
Data frame of goal events with timing
}
\description{
When match summary data isn't available in the pre-scraped cache,
we can derive goal events from the shooting data (which IS available).
This gives us goal timing for splint creation, though not substitution timing.
}
\keyword{internal}
