% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DIRECT_CORNER_X_MIN}
\alias{DIRECT_CORNER_X_MIN}
\alias{DIRECT_CORNER_Y_EDGE}
\title{Corner-flag box for untagged direct corners: x at or beyond 97 and within 4
of a touchline, Opta 0-100 coordinates (panna#277).}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
DIRECT_CORNER_X_MIN

DIRECT_CORNER_Y_EDGE
}
\description{
Corner-flag box for untagged direct corners: x at or beyond 97 and within 4
of a touchline, Opta 0-100 coordinates (panna#277).
}
\keyword{datasets}
\keyword{internal}
