% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_matrix.R
\name{.aggregate_player_minutes}
\alias{.aggregate_player_minutes}
\title{Aggregate player minutes and split into regular/replacement pools}
\usage{
.aggregate_player_minutes(players, splints, min_minutes)
}
\arguments{
\item{players}{Data frame of player appearances per splint}

\item{splints}{Data frame of splints with duration}

\item{min_minutes}{Minimum total minutes for inclusion as regular player}
}
\value{
List with player_minutes, replacement_player_ids, player_ids,
n_players, all_player_minutes
}
\description{
Aggregate player minutes and split into regular/replacement pools
}
\keyword{internal}
