% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_opta.R
\name{.calculate_opta_per90}
\alias{.calculate_opta_per90}
\title{Calculate per-90 rates for Opta player stats}
\usage{
.calculate_opta_per90(player_stats)
}
\arguments{
\item{player_stats}{Data frame with aggregated counting stats}
}
\value{
player_stats with per-90 columns added
}
\description{
Calculate per-90 rates for Opta player stats
}
\keyword{internal}
