% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_opta.R
\name{.clean_numeric_na}
\alias{.clean_numeric_na}
\title{Replace NA/Inf with 0 in numeric columns and log summary}
\usage{
.clean_numeric_na(df, check_inf = TRUE)
}
\arguments{
\item{df}{data.frame with numeric columns}

\item{check_inf}{Whether to also replace Inf values (default TRUE)}
}
\value{
df with NAs/Inf replaced
}
\description{
Replace NA/Inf with 0 in numeric columns and log summary
}
\keyword{internal}
