% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_skills.R
\name{.extract_end_year}
\alias{.extract_end_year}
\title{Extract season end year from season string}
\usage{
.extract_end_year(season)
}
\arguments{
\item{season}{Character vector of season strings}
}
\value{
Integer vector of end years
}
\description{
Extract season end year from season string
}
\keyword{internal}
