% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.get_col_warned}
\alias{.get_col_warned}
\title{Memoization environment for .get_col warnings}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.get_col_warned
}
\description{
Tracks which missing-column warnings have been emitted by \code{.get_col()}.
Each column name is stored as a key once warned, preventing duplicate warnings
within a session. To reset: \code{rm(list = ls(.get_col_warned), envir = .get_col_warned)}
}
\keyword{internal}
