% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{.get_defense_col_mapping}
\alias{.get_defense_col_mapping}
\title{Get defense stats column mapping}
\usage{
.get_defense_col_mapping()
}
\description{
Get defense stats column mapping
}
\keyword{internal}
