% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_opta.R
\name{.get_opta_col_mapping}
\alias{.get_opta_col_mapping}
\title{Get Opta column mapping (new_name = opta_column_name)}
\usage{
.get_opta_col_mapping()
}
\value{
Named character vector mapping panna names to Opta column names
}
\description{
Get Opta column mapping (new_name = opta_column_name)
}
\keyword{internal}
