% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{.get_passing_col_mapping}
\alias{.get_passing_col_mapping}
\title{Get passing stats column mapping}
\usage{
.get_passing_col_mapping()
}
\description{
Get passing stats column mapping
}
\keyword{internal}
