% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{.get_possession_col_mapping}
\alias{.get_possession_col_mapping}
\title{Get possession stats column mapping}
\usage{
.get_possession_col_mapping()
}
\description{
Get possession stats column mapping
}
\keyword{internal}
