% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{.get_summary_col_mapping}
\alias{.get_summary_col_mapping}
\title{Get summary stats column mapping}
\usage{
.get_summary_col_mapping()
}
\description{
Get summary stats column mapping
}
\keyword{internal}
