% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr.R
\name{.psv_position_group}
\alias{.psv_position_group}
\title{Bucket an Opta position label into the calibration's position groups}
\usage{
.psv_position_group(p)
}
\description{
Bucket an Opta position label into the calibration's position groups
}
\keyword{internal}
