% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_opta.R
\name{.rename_opta_columns}
\alias{.rename_opta_columns}
\title{Rename Opta columns to panna names using the standard mapping}
\usage{
.rename_opta_columns(dt)
}
\arguments{
\item{dt}{data.table or data.frame}
}
\value{
The input with columns renamed
}
\description{
Rename Opta columns to panna names using the standard mapping
}
\keyword{internal}
