% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_CONF_SPREAD}
\alias{ELO_CONF_SPREAD}
\title{Default conf_spread used by step 03 (passed through compute_match_elos)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_CONF_SPREAD
}
\description{
Default conf_spread used by step 03 (passed through compute_match_elos)
}
\keyword{internal}
