% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_DAVIDSON_NU}
\alias{ELO_DAVIDSON_NU}
\title{Optimized Davidson draw parameter (v6 = 0.89)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_DAVIDSON_NU
}
\description{
Optimized Davidson draw parameter (v6 = 0.89)
}
\keyword{internal}
