% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_DEFAULT_K}
\alias{ELO_DEFAULT_K}
\title{Default base K for any league not in \code{ELO_MATCH_TYPE_K}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_DEFAULT_K
}
\description{
Default base K for any league not in \code{ELO_MATCH_TYPE_K}
}
\keyword{internal}
