% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_HOME_ADV}
\alias{ELO_HOME_ADV}
\title{Optimized home-advantage value (Elo points; v6 = 88, v3 = 65)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_HOME_ADV
}
\description{
Optimized home-advantage value (Elo points; v6 = 88, v3 = 65)
}
\keyword{internal}
