% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_HOST_NAME_ALIASES}
\alias{ELO_HOST_NAME_ALIASES}
\title{Map Host-Name Aliases to Canonical Opta team_name}
\format{
An object of class \code{character} of length 6.
}
\usage{
ELO_HOST_NAME_ALIASES
}
\description{
Tournament season strings encode the host country (e.g., "2024 USA")
but the host may appear under a different spelling in opta_lineups
(here "United States"). This map normalises common aliases.
}
\keyword{internal}
