% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_DEFENSIVE_BOOST}
\alias{EPV_DEFENSIVE_BOOST}
\title{Defensive action credit boost multiplier}
\format{
Numeric value: 1.5
}
\usage{
EPV_DEFENSIVE_BOOST
}
\description{
Multiplier applied to defensive actions (clearance, interception,
tackle, ball_recovery) to better capture their value. EPV deltas
underestimate defensive contributions.
}
\keyword{internal}
