% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_OPP_PRIOR_GAMES}
\alias{EPV_OPP_PRIOR_GAMES}
\title{Opponent adjustment: prior games for shrinkage}
\format{
Numeric value: 2
}
\usage{
EPV_OPP_PRIOR_GAMES
}
\description{
Number of pseudo-games at league average for Bayesian shrinkage of
opponent profiles. Lower = faster response, higher = more stable.
2 is appropriate for within-season single-league data.
}
\keyword{internal}
