% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_POSITION_RAMP_X}
\alias{EPV_POSITION_RAMP_X}
\title{X-coordinate where position scaling reaches 1.0}
\format{
Numeric value: 40
}
\usage{
EPV_POSITION_RAMP_X
}
\description{
Passes from beyond this x-coordinate get full credit.
Below this, credit is scaled down toward EPV_POSITION_SCALE_MIN.
}
\keyword{internal}
