% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_TURNOVER_BLAME_SHARE}
\alias{EPV_TURNOVER_BLAME_SHARE}
\title{Blame share for turnovers, saved shots, and duels}
\format{
Numeric value: 0.5
}
\usage{
EPV_TURNOVER_BLAME_SHARE
}
\description{
When a turnover or duel results in negative EPV delta, this fraction
of blame goes to the actor (the rest to the receiver who gains).
}
\keyword{internal}
