% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_model.R
\name{extract_od_rapm_coefficients}
\alias{extract_od_rapm_coefficients}
\title{Extract offensive/defensive RAPM coefficients}
\usage{
extract_od_rapm_coefficients(model, lambda = "min")
}
\arguments{
\item{model}{Fitted RAPM model with O/D separated matrix}

\item{lambda}{Which lambda to use}
}
\value{
Data frame with offensive and defensive ratings
}
\description{
Gets separate O-RAPM and D-RAPM from a model fit on O/D matrix.
}
