% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_model.R
\name{extract_rapm_coefficients}
\alias{extract_rapm_coefficients}
\title{Extract RAPM coefficients}
\usage{
extract_rapm_coefficients(model, lambda = "min")
}
\arguments{
\item{model}{Fitted RAPM model from fit_rapm}

\item{lambda}{Which lambda to use ("min" or "1se" or numeric)}
}
\value{
Data frame with player ratings
}
\description{
Gets player ratings from a fitted RAPM model.
}
