% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{extract_season_from_url}
\alias{extract_season_from_url}
\title{Extract season from FBref match URL}
\usage{
extract_season_from_url(urls)
}
\arguments{
\item{urls}{Character vector of match URLs}
}
\value{
Character vector of seasons
}
\description{
Extract season from FBref match URL
}
\keyword{internal}
