% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{extract_spm_coefficients}
\alias{extract_spm_coefficients}
\title{Extract SPM coefficients}
\usage{
extract_spm_coefficients(model, lambda = "min")
}
\arguments{
\item{model}{Fitted SPM model from fit_spm_model}

\item{lambda}{Which lambda to use ("min" or "1se")}
}
\value{
Named vector of coefficients
}
\description{
Gets feature weights from fitted SPM model.
}
