% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_matrix.R
\name{filter_rapm_by_season}
\alias{filter_rapm_by_season}
\title{Filter RAPM data to specific seasons}
\usage{
filter_rapm_by_season(rapm_data, seasons, match_info)
}
\arguments{
\item{rapm_data}{List from prepare_rapm_data}

\item{seasons}{Character vector of seasons to include}

\item{match_info}{Match info data frame with season column}
}
\value{
Filtered rapm_data
}
\description{
Subsets RAPM data to include only splints from specified seasons.
}
