% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{find_components}
\alias{find_components}
\title{Find Connected Components}
\usage{
find_components(adj)
}
\arguments{
\item{adj}{Sparse adjacency matrix}
}
\value{
List with membership, sizes, n_components
}
\description{
Find Connected Components
}
\keyword{internal}
