% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{fit_panna_model}
\alias{fit_panna_model}
\title{Fit end-to-end panna model}
\usage{
fit_panna_model(
  splint_data,
  player_features,
  min_minutes = 180,
  lambda_prior = 1,
  validate = TRUE
)
}
\arguments{
\item{splint_data}{Combined splint data}

\item{player_features}{Player feature matrix}

\item{min_minutes}{Minimum minutes for inclusion}

\item{lambda_prior}{Regularization strength}

\item{validate}{Whether to run validation}
}
\value{
List with panna model and all intermediate results
}
\description{
Complete pipeline from splint data to panna ratings.
}
\examples{
\dontrun{
panna_model <- fit_panna_model(splint_data, player_features, min_minutes = 180)
head(get_panna_ratings(panna_model))
}

}
\keyword{internal}
