% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_direct.R
\name{get_all_competition_codes}
\alias{get_all_competition_codes}
\title{Get all supported competition codes}
\usage{
get_all_competition_codes()
}
\value{
Character vector of all competition codes
}
\description{
Get all supported competition codes
}
