% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{get_big5_leagues}
\alias{get_big5_leagues}
\title{Big 5 European league configurations}
\usage{
get_big5_leagues()
}
\value{
Data frame with league info
}
\description{
Big 5 European league configurations
}
