% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get cache directory path}
\usage{
get_cache_dir(create = TRUE)
}
\arguments{
\item{create}{Logical, whether to create if doesn't exist}
}
\value{
Path to cache directory
}
\description{
Get cache directory path
}
\keyword{internal}
