% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_utils.R
\name{get_cached_match_urls_legacy}
\alias{get_cached_match_urls_legacy}
\title{Get already-scraped match URLs from legacy cache}
\usage{
get_cached_match_urls_legacy(data_type)
}
\arguments{
\item{data_type}{Type of data to check}
}
\value{
Character vector of match URLs already in cache
}
\description{
Get already-scraped match URLs from legacy cache
}
\keyword{internal}
