% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat.R
\name{get_cached_understat_ids}
\alias{get_cached_understat_ids}
\title{Get cached Understat match IDs}
\usage{
get_cached_understat_ids(league, season)
}
\arguments{
\item{league}{League code}

\item{season}{Season string}
}
\value{
Character vector of cached understat_ids
}
\description{
Get cached Understat match IDs
}
