% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm_model.R
\name{get_covariate_effects}
\alias{get_covariate_effects}
\title{Get covariate effects from fitted model}
\usage{
get_covariate_effects(model, lambda = "min")
}
\arguments{
\item{model}{Fitted RAPM model from fit_rapm}

\item{lambda}{Which lambda to use}
}
\value{
Named vector of covariate coefficients
}
\description{
Extracts the coefficients for game state covariates.
}
