% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_direct.R
\name{get_domestic_cup_codes}
\alias{get_domestic_cup_codes}
\title{Get domestic cup codes}
\usage{
get_domestic_cup_codes()
}
\value{
Character vector of cup codes
}
\description{
Get domestic cup codes
}
