% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{get_first_half_stoppage}
\alias{get_first_half_stoppage}
\title{Get first-half stoppage time from events}
\usage{
get_first_half_stoppage(events)
}
\arguments{
\item{events}{Data frame with minute and optionally added_time columns}
}
\value{
Integer, first-half stoppage in minutes (0 if none)
}
\description{
Finds the maximum added_time for events at minute 45.
}
\keyword{internal}
