% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{get_first_half_stoppage_from_shots}
\alias{get_first_half_stoppage_from_shots}
\title{Get first-half stoppage from shots data}
\usage{
get_first_half_stoppage_from_shots(shots)
}
\arguments{
\item{shots}{Data frame with minute column (character, e.g., "45+3")}
}
\value{
Integer, first-half stoppage in minutes (0 if none)
}
\description{
Parses shot minute strings to find max added_time at minute 45.
}
\keyword{internal}
