% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{get_match_end_from_shots}
\alias{get_match_end_from_shots}
\title{Get match end from shots data}
\usage{
get_match_end_from_shots(shots, first_half_stoppage = 0L)
}
\arguments{
\item{shots}{Data frame with minute column (character, e.g., "90+4")}

\item{first_half_stoppage}{First half stoppage to offset second half}
}
\value{
Numeric match end minute (effective time)
}
\description{
Uses the last shot to determine match end time.
}
\keyword{internal}
