% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{get_panna_ratings}
\alias{get_panna_ratings}
\title{Get panna ratings from fitted model}
\usage{
get_panna_ratings(panna_model)
}
\arguments{
\item{panna_model}{Fitted panna model from \code{\link{fit_panna_model}}.}
}
\value{
Data frame with player panna ratings.
}
\description{
Extracts the final panna ratings from a fitted panna model object.
Requires a model fitted with \code{\link{fit_panna_model}}.
}
\details{
To generate ratings, run the Opta pipeline:
\code{source("data-raw/player-ratings-opta/run_pipeline_opta.R")}
}
\examples{
\dontrun{
panna_model <- fit_panna_model(splint_data, player_features)
ratings <- get_panna_ratings(panna_model)
head(ratings)
}

}
