% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{get_pitch_zone}
\alias{get_pitch_zone}
\title{Determine Pitch Zone}
\usage{
get_pitch_zone(x, y)
}
\arguments{
\item{x}{X coordinate (0-100)}

\item{y}{Y coordinate (0-100)}
}
\value{
Integer zone ID (1-18)
}
\description{
Assigns a zone ID based on pitch location. Uses a 3x6 grid (18 zones).
}
\examples{
get_pitch_zone(10, 50)   # Defensive third, center
get_pitch_zone(90, 20)   # Attacking third, left
}
