% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_direct.R
\name{get_seasons_since}
\alias{get_seasons_since}
\title{Get all seasons since 2017-2018}
\usage{
get_seasons_since(start_year = 2017)
}
\arguments{
\item{start_year}{First season start year (default 2017)}
}
\value{
Character vector of season strings
}
\description{
Returns a vector of season strings from 2017-2018 to current season.
}
