% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{get_source_archive_name}
\alias{get_source_archive_name}
\title{Get archive filename for data source}
\usage{
get_source_archive_name(source_type)
}
\arguments{
\item{source_type}{One of "fbref", "understat", "opta", or "all"}
}
\value{
Character archive filename (tar.gz)
}
\description{
Maps source type to tar.gz archive filename.
}
\keyword{internal}
