% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{get_source_tag}
\alias{get_source_tag}
\title{Get release tag for data source}
\usage{
get_source_tag(source_type)
}
\arguments{
\item{source_type}{One of "fbref", "understat", "opta", or "all"}
}
\value{
Character release tag name
}
\description{
Maps source type to GitHub release tag name.
}
\keyword{internal}
