% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{get_source_zipname}
\alias{get_source_zipname}
\title{Get zip filename for data source}
\usage{
get_source_zipname(source_type)
}
\arguments{
\item{source_type}{One of "fbref", "understat", "opta", or "all"}
}
\value{
Character zip filename
}
\description{
Maps source type to zip filename.
}
\keyword{internal}
