% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{get_splint_players}
\alias{get_splint_players}
\title{Get players on pitch for a splint}
\usage{
get_splint_players(splint_data, splint_id)
}
\arguments{
\item{splint_data}{Combined splint data from create_all_splints}

\item{splint_id}{Splint identifier}
}
\value{
List with home_players and away_players character vectors
}
\description{
Returns the player IDs for home and away teams in a splint.
}
\keyword{internal}
