% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_stat_columns}
\alias{get_stat_columns}
\title{Get column names for stat type}
\usage{
get_stat_columns(stat_type)
}
\arguments{
\item{stat_type}{Type of statistics}
}
\value{
Character vector of column names
}
\description{
Returns expected column names from pannadata/FBref data.
}
\keyword{internal}
