% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_direct.R
\name{get_uefa_cup_codes}
\alias{get_uefa_cup_codes}
\title{Get UEFA club competition codes}
\usage{
get_uefa_cup_codes()
}
\value{
Character vector of UEFA competition codes
}
\description{
Get UEFA club competition codes
}
