% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat_api.R
\name{get_understat_scripts}
\alias{get_understat_scripts}
\title{Get all script tags from Understat page}
\usage{
get_understat_scripts(page)
}
\arguments{
\item{page}{Parsed HTML document}
}
\value{
Character vector of script contents
}
\description{
Extracts the text content of all script tags.
}
\keyword{internal}
