% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_competitions.R
\name{get_understat_seasons}
\alias{get_understat_seasons}
\title{Get available Understat seasons}
\usage{
get_understat_seasons(start_year = 2014)
}
\arguments{
\item{start_year}{First season year (default 2014)}
}
\value{
Character vector of season years
}
\description{
Returns a vector of available seasons for Understat data.
Understat has data from 2014 onwards.
}
\examples{
\dontrun{
get_understat_seasons(2020)
}
}
\keyword{internal}
