% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat.R
\name{get_understat_session}
\alias{get_understat_session}
\title{Get or create Understat session}
\usage{
get_understat_session(reset = FALSE)
}
\arguments{
\item{reset}{If TRUE, creates a new session (default FALSE)}
}
\value{
httr handle object
}
\description{
Returns a persistent httr session handle that maintains cookies.
}
