% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_competitions.R
\name{get_understat_slug}
\alias{get_understat_slug}
\title{Get Understat URL slug}
\usage{
get_understat_slug(code)
}
\arguments{
\item{code}{Character, the league code (e.g., "ENG", "RUS")}
}
\value{
Character, the Understat URL slug
}
\description{
Looks up the Understat URL slug for a given league code.
}
\examples{
\dontrun{
get_understat_slug("RUS")  # "RFPL"
}
}
\keyword{internal}
