% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HALFTIME_MINUTE}
\alias{HALFTIME_MINUTE}
\title{Minute marking halftime (end of first half)}
\format{
Integer value: 45
}
\usage{
HALFTIME_MINUTE
}
\description{
The minute at which the first half ends. Used for detecting
first-half stoppage time events.
}
\examples{
HALFTIME_MINUTE
}
\seealso{
Other constants: 
\code{\link{BETA_PRIOR_ALPHA}},
\code{\link{CHAIN_TIME_GAP_SECONDS}},
\code{\link{CONFIDENCE_LEVEL}},
\code{\link{MINUTES_PER_MATCH}},
\code{\link{MIN_90S_PSR_LEADERBOARD}},
\code{\link{MIN_GAMES_FOR_PADDING}},
\code{\link{MIN_MINUTES_FEATURES}},
\code{\link{MIN_MINUTES_RAPM}},
\code{\link{MIN_MINUTES_SPM}},
\code{\link{MIN_SEQUENCES_PER_MATCH}},
\code{\link{MIN_SHOTS_FOR_FINISHING}},
\code{\link{MIN_WEIGHT_DURATION}},
\code{\link{PENALTY_SHOOTOUT_CONVERSION}},
\code{\link{PENALTY_XG}},
\code{\link{PLAYERS_PER_TEAM}},
\code{\link{SIX_YARD_X_MIN}},
\code{\link{SIX_YARD_Y_MAX}},
\code{\link{SIX_YARD_Y_MIN}},
\code{\link{TOUCHES_PER_SEQUENCE}},
\code{\link{XG_MAX}},
\code{\link{XG_MIN}}
}
\concept{constants}
\keyword{datasets}
