% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{is_in_final_third}
\alias{is_in_final_third}
\title{Check if Coordinates are in Final Third}
\usage{
is_in_final_third(x, attacking = TRUE)
}
\arguments{
\item{x}{X coordinate (0-100)}

\item{attacking}{If TRUE, checks attacking final third (x > 67)}
}
\value{
Logical indicating if in final third
}
\description{
Check if Coordinates are in Final Third
}
