% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{is_in_penalty_area}
\alias{is_in_penalty_area}
\title{Check if Coordinates are in Penalty Area}
\usage{
is_in_penalty_area(x, y, attacking = TRUE)
}
\arguments{
\item{x}{X coordinate (0-100)}

\item{y}{Y coordinate (0-100)}

\item{attacking}{If TRUE, checks attacking penalty area (x > 83)}
}
\value{
Logical indicating if in penalty area
}
\description{
Check if Coordinates are in Penalty Area
}
