% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat.R
\name{is_understat_cached}
\alias{is_understat_cached}
\title{Check if Understat match is cached}
\usage{
is_understat_cached(league, season, understat_id)
}
\arguments{
\item{league}{League code}

\item{season}{Season string}

\item{understat_id}{Understat match ID}
}
\value{
Logical - TRUE if match has been scraped
}
\description{
Check if Understat match is cached
}
\keyword{internal}
