% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_competitions.R
\name{is_understat_league}
\alias{is_understat_league}
\title{Check if league is available on Understat}
\usage{
is_understat_league(code)
}
\arguments{
\item{code}{Character, the league code}
}
\value{
Logical, TRUE if league is available on Understat
}
\description{
Check if league is available on Understat
}
\examples{
\dontrun{
is_understat_league("POR")
}
}
\keyword{internal}
