% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_competitions.R
\name{list_competitions}
\alias{list_competitions}
\title{List competitions by type}
\usage{
list_competitions(type = "all")
}
\arguments{
\item{type}{Character, one of "league", "cup", "european", "national_team", or "all"}
}
\value{
Character vector of competition codes
}
\description{
Returns competition codes filtered by type.
}
\examples{
list_competitions("league")
list_competitions("national_team")
}
