% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_competitions.R
\name{list_understat_competitions}
\alias{list_understat_competitions}
\title{List Understat competitions}
\usage{
list_understat_competitions()
}
\value{
Character vector of competition codes
}
\description{
Returns all available Understat competition codes.
}
\examples{
list_understat_competitions()
}
