% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat.R
\name{load_understat_table}
\alias{load_understat_table}
\title{Load Understat data from parquet cache}
\usage{
load_understat_table(league, season, understat_id = NULL, table_type)
}
\arguments{
\item{league}{League code}

\item{season}{Season string}

\item{understat_id}{Optional specific match ID to load}

\item{table_type}{Type of table}
}
\value{
Data frame or NULL if not cached
}
\description{
Load Understat data from parquet cache
}
\keyword{internal}
