% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{map_opta_action_type}
\alias{map_opta_action_type}
\title{Map Opta Type ID to SPADL Action Type}
\usage{
map_opta_action_type(type_id, qualifiers = NULL)
}
\arguments{
\item{type_id}{Vector of Opta type_id values}

\item{qualifiers}{Parsed qualifier information (optional)}
}
\value{
Character vector of SPADL action types
}
\description{
Map Opta Type ID to SPADL Action Type
}
\keyword{internal}
