% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{map_opta_bodypart}
\alias{map_opta_bodypart}
\title{Map Opta Events to Body Part}
\usage{
map_opta_bodypart(type_id, qualifiers = NULL)
}
\arguments{
\item{type_id}{Vector of Opta type_id values}

\item{qualifiers}{Parsed qualifier information (optional)}
}
\value{
Character vector of body parts: "foot", "head", or "other"
}
\description{
Map Opta Events to Body Part
}
\keyword{internal}
