% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{merge_processed_data}
\alias{merge_processed_data}
\title{Merge all processed data}
\usage{
merge_processed_data(processed_data)
}
\arguments{
\item{processed_data}{List of processed data frames}
}
\value{
List with merged data frames
}
\description{
Combines all processed data sources into unified datasets.
}
\keyword{internal}
