% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna-package.R
\docType{package}
\name{panna-package}
\alias{panna}
\alias{panna-package}
\title{panna: Player Rating System for Football Using RAPM, SPM, and EPV Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements player ratings for football (soccer) from 'Opta' event and tracking-adjacent data. Combines Regularized Adjusted Plus-Minus ('RAPM') with Statistical Plus-Minus ('SPM') as a Bayesian prior, analysing time segments with constant lineups ('splints') to estimate each player's contribution to expected goals ('xG') differential. Action-level Expected Possession Value ('EPV') and win probability added ('WPA') models support decay-weighted rating composites (EPR, PSR, Panna, Piero) plus match prediction and tournament simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/peteowen1/panna}
  \item \url{https://peteowen1.github.io/panna/}
  \item Report bugs at \url{https://github.com/peteowen1/panna/issues}
}

}
\author{
\strong{Maintainer}: Pete Owen \email{pete.owen1@hotmail.co.uk} [copyright holder]

}
\keyword{internal}
