% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PANNA_DOMESTIC_LEAGUES}
\alias{PANNA_DOMESTIC_LEAGUES}
\title{Domestic-only competitions (for league-offset attribution)}
\format{
Character vector of domestic competition codes
}
\usage{
PANNA_DOMESTIC_LEAGUES
}
\description{
The subset of \code{PANNA_RATING_LEAGUES} that is a domestic league. Used
when attributing a player-season to "the league he plays in", which must be
a domestic competition -- a cross-league cup is where leagues MEET, not a
league a player belongs to.
}
\details{

}
\keyword{internal}
