% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PANNA_PSR_WEIGHT}
\alias{PANNA_PSR_WEIGHT}
\title{PSR weight in combined Panna Value rating}
\format{
Numeric value: 0.5
}
\usage{
PANNA_PSR_WEIGHT
}
\description{
PSR weight in combined Panna Value rating
}
\seealso{
Other epr: 
\code{\link{PANNA_EPR_WEIGHT}},
\code{\link{aggregate_player_game_epv}()},
\code{\link{build_player_game_ratings}()},
\code{\link{calculate_epr}()},
\code{\link{calculate_epr_batch}()},
\code{\link{calculate_epr_regression}()},
\code{\link{optimize_epr_decay}()},
\code{\link{player_value}()}
}
\concept{epr}
\keyword{datasets}
