% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{parse_event_minute}
\alias{parse_event_minute}
\title{Parse event minute from time string}
\usage{
parse_event_minute(time_str)
}
\arguments{
\item{time_str}{Character vector of event times}
}
\value{
Numeric vector of minutes
}
\description{
Handles formats like "45+2" (45 plus 2 minutes stoppage time).
}
\keyword{internal}
