% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{pb_sync_data}
\alias{pb_sync_data}
\title{Sync local data with GitHub Releases}
\usage{
pb_sync_data(
  repo = "peteowen/pannadata",
  tag = "data",
  source = NULL,
  table_types = NULL
)
}
\arguments{
\item{repo}{GitHub repository in "owner/repo" format}

\item{tag}{Release tag (default: "data")}

\item{source}{Source directory (default: pannadata_dir())}

\item{table_types}{Which table types to sync (default: all)}
}
\value{
Number of new files uploaded
}
\description{
Uploads any local files that aren't in the release yet.
This is a more targeted upload than pb_upload_data.
}
