% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{per_90}
\alias{per_90}
\title{Calculate minutes per 90}
\usage{
per_90(stat, minutes)
}
\arguments{
\item{stat}{Statistic value (counting stat)}

\item{minutes}{Minutes played}
}
\value{
Statistic per 90 minutes
}
\description{
Calculate minutes per 90
}
\seealso{
Other utilities: 
\code{\link{safe_divide}()},
\code{\link{validate_dataframe}()},
\code{\link{validate_seasons}()}
}
\concept{utilities}
