% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{predict_epv_probs}
\alias{predict_epv_probs}
\title{Predict EPV Values}
\usage{
predict_epv_probs(model, features)
}
\arguments{
\item{model}{EPV model from fit_epv_model()}

\item{features}{EPV features}
}
\value{
For "goal": data frame with p_team_scores, p_opponent_scores, p_nobody_scores
For "xg": data frame with expected_xg
}
\description{
Gets predictions from EPV model. For "goal" method returns multinomial
probabilities, for "xg" method returns expected next xG value.
}
\keyword{internal}
