% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{prepare_od_scatter_data}
\alias{prepare_od_scatter_data}
\title{Create O/D scatter plot data}
\usage{
prepare_od_scatter_data(ratings)
}
\arguments{
\item{ratings}{O/D ratings data frame}
}
\value{
Data frame ready for ggplot2
}
\description{
Prepares data for visualizing offensive vs defensive ratings.
}
\keyword{internal}
