% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{prepare_spm_regression_data}
\alias{prepare_spm_regression_data}
\title{Prepare SPM regression data}
\usage{
prepare_spm_regression_data(player_features, rapm_ratings)
}
\arguments{
\item{player_features}{Data frame from create_player_feature_matrix}

\item{rapm_ratings}{Data frame from extract_rapm_coefficients}
}
\value{
Data frame ready for SPM regression
}
\description{
Joins player features with RAPM ratings for SPM model fitting.
}
