% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_match_events}
\alias{process_match_events}
\title{Process match events}
\usage{
process_match_events(events, results)
}
\arguments{
\item{events}{Data frame of match events}

\item{results}{Processed match results for match IDs}
}
\value{
Cleaned data frame with match events
}
\description{
Cleans event data (goals, substitutions) with timing.
}
\keyword{internal}
