% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_shooting_data}
\alias{process_shooting_data}
\title{Process shooting data}
\usage{
process_shooting_data(shooting, results)
}
\arguments{
\item{shooting}{Data frame from scrape_pl_match_shooting}

\item{results}{Processed match results for match IDs}
}
\value{
Cleaned data frame with shot data
}
\description{
Cleans shot-level data with xG values.
}
