% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress_msg}
\alias{progress_msg}
\title{Print progress message}
\usage{
progress_msg(msg, verbose = TRUE)
}
\arguments{
\item{msg}{Message to print}

\item{verbose}{Whether to print}
}
\value{
NULL (invisibly)
}
\description{
Print progress message
}
\keyword{internal}
