% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REGULATION_SECONDS}
\alias{REGULATION_SECONDS}
\title{Regulation seconds (90 minutes)}
\format{
Numeric value: 5400
}
\usage{
REGULATION_SECONDS
}
\description{
Duration of regulation time in seconds. Win-probability time features use
this denominator for matches that did NOT reach extra time.
}
\keyword{internal}
