% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_utils.R
\name{reset_fbref_session}
\alias{reset_fbref_session}
\title{Reset FBref session}
\usage{
reset_fbref_session()
}
\description{
Clears cookies and creates fresh session. Use after changing VPN/IP.
}
\keyword{internal}
