% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{save_cached_data}
\alias{save_cached_data}
\title{Save data to cache}
\usage{
save_cached_data(data, data_type)
}
\arguments{
\item{data}{Data frame to cache}

\item{data_type}{Type of data}
}
\description{
Save data to cache
}
\keyword{internal}
