% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duel_model.R
\name{save_duel_model}
\alias{save_duel_model}
\title{Save the xDuel model}
\usage{
save_duel_model(duel_model, path = NULL)
}
\arguments{
\item{duel_model}{Fitted model.}

\item{path}{Optional path. Defaults to pannadata opta/models/duel_model.rds.}
}
\value{
Invisibly, the path.
}
\description{
Save the xDuel model
}
\seealso{
Other xduel: 
\code{\link{compute_all_duel_preps}()},
\code{\link{compute_duel_woe}()},
\code{\link{fit_duel_model}()},
\code{\link{load_duel_model}()},
\code{\link{predict_duel}()},
\code{\link{prepare_duels_from_events}()}
}
\concept{xduel}
