% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{save_epv_model}
\alias{save_epv_model}
\title{Save EPV Model}
\usage{
save_epv_model(epv_model, path = NULL)
}
\arguments{
\item{epv_model}{EPV model from fit_epv_model()}

\item{path}{Directory to save model. If NULL, uses pannadata/data/opta/models/}
}
\value{
Invisibly returns path
}
\description{
Saves trained EPV model to disk.
}
\seealso{
Other epv: 
\code{\link{aggregate_player_xmetrics}()},
\code{\link{assign_epv_credit}()},
\code{\link{convert_opta_to_spadl}()},
\code{\link{enrich_match_stats_with_xmetrics}()},
\code{\link{fit_epv_model}()},
\code{\link{fit_xg_model}()},
\code{\link{fit_xgot_model}()},
\code{\link{fit_xpass_model}()},
\code{\link{get_or_build_spadl}()},
\code{\link{load_epv_model}()},
\code{\link{load_xg_model}()},
\code{\link{load_xgot_model}()},
\code{\link{load_xpass_model}()},
\code{\link{pb_download_epv_models}()},
\code{\link{predict_xg}()},
\code{\link{predict_xgot}()},
\code{\link{predict_xpass}()}
}
\concept{epv}
