% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_cache.R
\name{save_match_table}
\alias{save_match_table}
\title{Save match table to cache}
\usage{
save_match_table(data, league, season, fbref_id, table_type)
}
\arguments{
\item{data}{Data frame to save}

\item{league}{League code}

\item{season}{Season string}

\item{fbref_id}{FBref match ID}

\item{table_type}{Type of table}
}
\value{
Invisible path to saved file
}
\description{
Saves to hierarchical path: \code{\{table_type\}/\{league\}/\{season\}/\{id\}.rds}
}
\keyword{internal}
