% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_understat.R
\name{save_understat_manifest}
\alias{save_understat_manifest}
\title{Save Understat manifest to parquet file}
\usage{
save_understat_manifest(manifest, path)
}
\arguments{
\item{manifest}{Data frame with manifest data}

\item{path}{Path to save manifest parquet file}
}
\value{
Invisible path to saved file
}
\description{
Save Understat manifest to parquet file
}
