% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_model.R
\name{save_wp_model}
\alias{save_wp_model}
\title{Save WP model}
\usage{
save_wp_model(wp_model, path = NULL)
}
\arguments{
\item{wp_model}{WP model from \code{\link{train_wp_model}}.}

\item{path}{Directory to save. If NULL, uses \code{pannadata/data/opta/models/}.}
}
\value{
Invisibly returns the file path.
}
\description{
Save WP model
}
\seealso{
Other win probability: 
\code{\link{add_wp_vars}()},
\code{\link{aggregate_player_game_wpa}()},
\code{\link{assign_wpa_credit}()},
\code{\link{create_wp_features}()},
\code{\link{load_wp_model}()},
\code{\link{predict_wp}()},
\code{\link{train_wp_model}()}
}
\concept{win probability}
