% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{save_xpass_model}
\alias{save_xpass_model}
\title{Save xPass Model}
\usage{
save_xpass_model(xpass_model, path = NULL)
}
\arguments{
\item{xpass_model}{Fitted xPass model}

\item{path}{Path to save. If NULL, saves to pannadata/opta/models/}
}
\value{
Invisibly returns the path
}
\description{
Saves trained xPass model to RDS file.
}
\examples{
\dontrun{
xpass_model <- fit_xpass_model(pass_features)
save_xpass_model(xpass_model)
save_xpass_model(xpass_model, path = "models/my_xpass_model.rds")
}
}
\keyword{internal}
