% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{scrape_league_data}
\alias{scrape_league_data}
\title{Scrape data for a single league}
\usage{
scrape_league_data(country, tier, seasons, delay = 3, verbose = TRUE)
}
\arguments{
\item{country}{Country code (ENG, ESP, GER, ITA, FRA)}

\item{tier}{League tier (usually "1st")}

\item{seasons}{Character vector of seasons}

\item{delay}{Seconds between requests}

\item{verbose}{Print progress}
}
\value{
List of data frames for this league
}
\description{
Internal function to collect all data for one league.
}
\keyword{internal}
