% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{scrape_pl_player_stats}
\alias{scrape_pl_player_stats}
\title{Scrape season player stats}
\usage{
scrape_pl_player_stats(seasons, stat_type = "standard")
}
\arguments{
\item{seasons}{Character vector of seasons}

\item{stat_type}{Type of stats: "standard", "shooting", "passing",
"passing_types", "gca", "defense", "possession", "misc"}
}
\value{
Data frame of player season stats
}
\description{
Fetches aggregated player statistics for full seasons.
}
\examples{
\dontrun{
stats <- scrape_pl_player_stats("2023-2024", "standard")
}
}
