% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPADL_CACHE_DIR}
\alias{SPADL_CACHE_DIR}
\title{Default SPADL cache directory}
\format{
Character value
}
\usage{
SPADL_CACHE_DIR
}
\description{
Relative path (from panna/ root) to the SPADL conversion cache shared
between the EPV/xMetrics pipeline and the Opta RAPM pipeline.
}
\keyword{internal}
