% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_min_minutes}
\alias{validate_min_minutes}
\title{Validate min_minutes parameter}
\usage{
validate_min_minutes(min_minutes)
}
\arguments{
\item{min_minutes}{Value to validate}
}
\description{
Checks that min_minutes is a single non-negative number.
}
\keyword{internal}
