% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{validate_predictive_power}
\alias{validate_predictive_power}
\title{Validate predictive power}
\usage{
validate_predictive_power(panna_ratings, next_season_rapm)
}
\arguments{
\item{panna_ratings}{Panna ratings from season N}

\item{next_season_rapm}{RAPM from season N+1}
}
\value{
Validation metrics
}
\description{
Tests panna ratings against next season performance.
}
\examples{
\dontrun{
metrics <- validate_predictive_power(panna_ratings_2023, rapm_ratings_2024)
metrics$correlation
metrics$rmse
}

}
\keyword{internal}
