% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versebus.R
\name{vb_producer_info}
\alias{vb_producer_info}
\title{Producer identity block for the manifest (GHA env or local)}
\usage{
vb_producer_info()
}
\description{
Producer identity block for the manifest (GHA env or local)
}
\seealso{
Other versebus: 
\code{\link{vb_asset_entry}()},
\code{\link{vb_atomic_write}()},
\code{\link{vb_cache_validate}()},
\code{\link{vb_classify_error}()},
\code{\link{vb_confirm_absent}()},
\code{\link{vb_download}()},
\code{\link{vb_generation}()},
\code{\link{vb_guard_accumulate}()},
\code{\link{vb_list_assets}()},
\code{\link{vb_publish}()},
\code{\link{vb_read_manifest}()},
\code{\link{vb_read_prev_manifest}()},
\code{\link{vb_sha256}()},
\code{\link{vb_write_manifest}()}
}
\concept{versebus}
\keyword{internal}
