% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versebus.R
\name{vb_sha256}
\alias{vb_sha256}
\title{sha256 of a file}
\usage{
vb_sha256(path)
}
\description{
sha256 of a file
}
\seealso{
Other versebus: 
\code{\link{vb_asset_entry}()},
\code{\link{vb_atomic_write}()},
\code{\link{vb_cache_validate}()},
\code{\link{vb_classify_error}()},
\code{\link{vb_confirm_absent}()},
\code{\link{vb_download}()},
\code{\link{vb_generation}()},
\code{\link{vb_guard_accumulate}()},
\code{\link{vb_list_assets}()},
\code{\link{vb_producer_info}()},
\code{\link{vb_publish}()},
\code{\link{vb_read_manifest}()},
\code{\link{vb_read_prev_manifest}()},
\code{\link{vb_write_manifest}()}
}
\concept{versebus}
\keyword{internal}
