% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WC2026_LEAGUE}
\alias{WC2026_LEAGUE}
\title{League code for the WC 2026 tournament}
\format{
An object of class \code{character} of length 1.
}
\usage{
WC2026_LEAGUE
}
\description{
League code for the WC 2026 tournament
}
\keyword{internal}
