% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WC2026_SEASON_LABEL}
\alias{WC2026_SEASON_LABEL}
\title{Season label for the WC 2026 tournament (as it appears in Opta fixtures)}
\format{
An object of class \code{character} of length 1.
}
\usage{
WC2026_SEASON_LABEL
}
\description{
Season label for the WC 2026 tournament (as it appears in Opta fixtures)
}
\keyword{internal}
