% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WP_DRAW_VALUE}
\alias{WP_DRAW_VALUE}
\title{WP draw value}
\format{
Numeric value: 0.5
}
\usage{
WP_DRAW_VALUE
}
\description{
A draw is worth 0.5 in WP terms (1 of 3 league points).
}
\keyword{internal}
