% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WPA_ACTOR_SHARE}
\alias{WPA_ACTOR_SHARE}
\title{WPA actor share}
\format{
Numeric value: 0.5
}
\usage{
WPA_ACTOR_SHARE
}
\description{
Fraction of WPA credited to the acting player (remainder to receiver).
}
\keyword{internal}
