% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_model.R
\name{add_xg_to_spadl}
\alias{add_xg_to_spadl}
\title{Add xG to SPADL Actions}
\usage{
add_xg_to_spadl(spadl_actions, xg_model)
}
\arguments{
\item{spadl_actions}{SPADL actions data frame}

\item{xg_model}{Fitted xG model}
}
\value{
SPADL actions with xg column added for shots
}
\description{
Adds xG predictions to shot actions in SPADL data.
}
\keyword{internal}
