% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BETA_PRIOR_ALPHA}
\alias{BETA_PRIOR_ALPHA}
\title{Beta prior alpha for finishing modifier (shrinkage toward 1.0)}
\format{
Integer value: 5
}
\usage{
BETA_PRIOR_ALPHA
}
\description{
Pseudocount added to both goals and xG when calculating finishing modifier.
Formula: (goals + BETA_PRIOR_ALPHA) / (xG + BETA_PRIOR_ALPHA)
This shrinks extreme values toward 1.0.
}
\examples{
BETA_PRIOR_ALPHA
}
\keyword{datasets}
