% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{calculate_action_epv}
\alias{calculate_action_epv}
\title{Calculate Action EPV Values}
\usage{
calculate_action_epv(spadl_actions, features, epv_model)
}
\arguments{
\item{spadl_actions}{SPADL actions data frame (with team_id, possession_change)}

\item{features}{EPV features from create_epv_features()}

\item{epv_model}{Fitted EPV model from fit_epv_model()}
}
\value{
SPADL actions with EPV columns added:
\itemize{
\item epv: EPV at this state
\item epv_delta: Change in EPV (with perspective handling)
\item For "goal" method: p_team_scores, p_opponent_scores, p_nobody_scores
\item For "xg" method: expected_xg
}
}
\description{
Computes EPV for each action. Supports two methods:
\itemize{
\item "goal": EPV = P(team_scores) - P(opponent_scores), bounded -1 to +1
\item "xg": EPV = expected xG of next shot (already signed)
}
}
\details{
Handles possession changes by flipping perspective:
\itemize{
\item Same team: delta = EPV_after - EPV_before
\item Team change: delta = (-EPV_after) - EPV_before
}
}
