% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{calculate_spm_ratings}
\alias{calculate_spm_ratings}
\title{Calculate SPM ratings for all players}
\usage{
calculate_spm_ratings(player_features, spm_model, lambda = "min")
}
\arguments{
\item{player_features}{Data frame of player features}

\item{spm_model}{Fitted SPM model}

\item{lambda}{Which lambda to use}
}
\value{
Data frame with SPM ratings
}
\description{
Applies SPM model to predict RAPM for all players with features.
}
\examples{
\dontrun{
spm_model <- fit_spm_model(spm_data)
ratings <- calculate_spm_ratings(player_features, spm_model)
head(ratings)
}

}
