% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{categorize_player_profile}
\alias{categorize_player_profile}
\title{Categorize player profile}
\usage{
categorize_player_profile(o_rating, d_rating)
}
\arguments{
\item{o_rating}{Offensive rating}

\item{d_rating}{Defensive rating}
}
\value{
Character vector of player types
}
\description{
Classifies players based on their O/D rating balance.
}
\examples{
\dontrun{
categorize_player_profile(0.8, 0.2)
categorize_player_profile(c(0.6, 0.3, 0.5), c(0.4, 0.7, 0.5))
}
}
