% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAIN_TIME_GAP_SECONDS}
\alias{CHAIN_TIME_GAP_SECONDS}
\title{Time gap threshold for chain breaks (seconds)}
\format{
Integer value: 30
}
\usage{
CHAIN_TIME_GAP_SECONDS
}
\description{
Maximum gap in seconds between consecutive actions before
a new possession chain is started.
}
\examples{
CHAIN_TIME_GAP_SECONDS
}
\keyword{datasets}
