% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{compare_panna_rapm_spm}
\alias{compare_panna_rapm_spm}
\title{Compare panna, RAPM, and SPM ratings}
\usage{
compare_panna_rapm_spm(panna_ratings, rapm_ratings, spm_ratings)
}
\arguments{
\item{panna_ratings}{Data frame of panna ratings}

\item{rapm_ratings}{Data frame of RAPM ratings}

\item{spm_ratings}{Data frame of SPM ratings}
}
\value{
Data frame comparing all rating types
}
\description{
Analyzes differences between the three rating types.
}
\examples{
\dontrun{
comparison <- compare_panna_rapm_spm(panna_ratings, rapm_ratings, spm_ratings)
head(comparison)
}

}
