% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONFIDENCE_LEVEL}
\alias{CONFIDENCE_LEVEL}
\title{Default confidence level for statistical intervals}
\format{
Numeric value: 0.95
}
\usage{
CONFIDENCE_LEVEL
}
\description{
Standard confidence level (95\%) for confidence intervals and
hypothesis tests throughout the package.
}
\examples{
CONFIDENCE_LEVEL
}
\keyword{datasets}
