% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint_creation.R
\name{create_all_splints}
\alias{create_all_splints}
\title{Create splints for all matches}
\usage{
create_all_splints(processed_data, include_goals = TRUE, verbose = TRUE)
}
\arguments{
\item{processed_data}{List of processed data from process_all_data}

\item{include_goals}{Whether to create splints at goal times}

\item{verbose}{Print progress messages}
}
\value{
List with combined splint data
}
\description{
Generates splint data for an entire dataset.
Uses data.table for fast pre-splitting by match_id.
}
\examples{
\dontrun{
processed_data <- process_all_data("ENG", "2024-2025")
splints <- create_all_splints(processed_data)
splints <- create_all_splints(processed_data, include_goals = FALSE, verbose = FALSE)
}
}
