% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.get_col_warned}
\alias{.get_col_warned}
\title{Memoization environment for .get_col warnings}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.get_col_warned
}
\description{
Memoization environment for .get_col warnings
}
\keyword{internal}
