% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calibration.R
\docType{data}
\name{ELO_CROSS_CONF_MULT}
\alias{ELO_CROSS_CONF_MULT}
\title{Default cross-confederation K multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_CROSS_CONF_MULT
}
\description{
Multiply K by this factor when the two teams are from different
confederations. Optimized to 1.5 (modest) -- the conf priors below
do most of the cross-pool calibration work, so the multiplier only
needs to be moderate. 1.0 disables.
}
\keyword{internal}
