% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{fit_epv_scoring_model}
\alias{fit_epv_scoring_model}
\title{Deprecated: Fit EPV Scoring Model}
\usage{
fit_epv_scoring_model(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=fit_epv_model]{fit_epv_model()}}}
}
\value{
An EPV model object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=fit_epv_model]{fit_epv_model()}} instead.
The unified EPV model now handles both scoring and conceding predictions
using a multinomial approach.
}
\examples{
\dontrun{
# Deprecated: use fit_epv_model() instead
model <- fit_epv_scoring_model(features, labels)
}

}
\seealso{
\code{\link[=fit_epv_model]{fit_epv_model()}}
}
