% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{generate_panna_report}
\alias{generate_panna_report}
\title{Generate prediction report}
\usage{
generate_panna_report(panna_model, top_n = 10)
}
\arguments{
\item{panna_model}{Fitted panna model}

\item{top_n}{Number of top/bottom players to show}
}
\value{
Character string report
}
\description{
Creates a formatted summary of panna ratings.
}
\examples{
\dontrun{
panna_model <- fit_panna_model(splint_data, player_features)
report <- generate_panna_report(panna_model, top_n = 15)
cat(report)
}

}
