% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fbref_data.R
\name{get_cached_match_urls}
\alias{get_cached_match_urls}
\title{Get already-scraped match URLs from cache}
\usage{
get_cached_match_urls(data_type)
}
\arguments{
\item{data_type}{Type of data to check}
}
\value{
Character vector of match URLs already in cache
}
\description{
Get already-scraped match URLs from cache
}
\keyword{internal}
