% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_competitions.R
\name{get_fbref_comp_id}
\alias{get_fbref_comp_id}
\title{Get FBref competition ID}
\usage{
get_fbref_comp_id(code)
}
\arguments{
\item{code}{Character, the league code (e.g., "ENG", "WC")}
}
\value{
Integer, the FBref competition ID
}
\description{
Looks up the FBref competition ID for a given league code.
}
\examples{
get_fbref_comp_id("ENG")  # 9
get_fbref_comp_id("WC")   # 1
}
