% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{get_panna_ratings}
\alias{get_panna_ratings}
\title{Get panna ratings from fitted model}
\usage{
get_panna_ratings(panna_model)
}
\arguments{
\item{panna_model}{Fitted panna model from fit_panna_model}
}
\value{
Data frame with player panna ratings
}
\description{
Gets the final panna ratings from a fitted panna model.
}
