% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_competitions.R
\name{get_seasons_since}
\alias{get_seasons_since}
\title{Get all seasons since a start year}
\usage{
get_seasons_since(start_year = 2017)
}
\arguments{
\item{start_year}{First season start year (default 2017)}
}
\value{
Character vector of season strings (e.g., "2017-2018", "2018-2019", ...)
}
\description{
Returns a vector of season strings from start_year to current season.
}
\examples{
get_seasons_since(2020)
get_seasons_since()  # From 2017
}
