% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_model.R
\name{get_spm_feature_importance}
\alias{get_spm_feature_importance}
\title{Get top SPM feature importance}
\usage{
get_spm_feature_importance(model, n = 10, lambda = "min")
}
\arguments{
\item{model}{Fitted SPM model}

\item{n}{Number of top features to return}

\item{lambda}{Which lambda to use}
}
\value{
Data frame of top features by absolute coefficient
}
\description{
Identifies the most important features in the SPM model.
}
\examples{
\dontrun{
spm_model <- fit_spm_model(spm_data)
top_features <- get_spm_feature_importance(spm_model, n = 15)
top_features
}

}
