% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{get_top_defensive}
\alias{get_top_defensive}
\title{Get top defensive players}
\usage{
get_top_defensive(ratings, n = 10)
}
\arguments{
\item{ratings}{O/D ratings data frame}

\item{n}{Number of players}
}
\value{
Data frame of top defensive players
}
\description{
Returns players with highest defensive contribution.
}
