% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{get_top_offensive}
\alias{get_top_offensive}
\title{Get top offensive players}
\usage{
get_top_offensive(ratings, n = 10)
}
\arguments{
\item{ratings}{O/D ratings data frame}

\item{n}{Number of players}
}
\value{
Data frame of top offensive players
}
\description{
Returns players with highest offensive contribution.
}
