% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HALFTIME_MINUTE}
\alias{HALFTIME_MINUTE}
\title{Minute marking halftime (end of first half)}
\format{
Integer value: 45
}
\usage{
HALFTIME_MINUTE
}
\description{
The minute at which the first half ends. Used for detecting
first-half stoppage time events.
}
\examples{
HALFTIME_MINUTE
}
\keyword{datasets}
