% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opta_loaders.R
\name{list_opta_seasons}
\alias{list_opta_seasons}
\title{List Available Opta Seasons}
\usage{
list_opta_seasons(league)
}
\arguments{
\item{league}{League code (e.g., "ENG", "EPL", "ESP", "La_Liga").}
}
\value{
Character vector of available seasons.
}
\description{
Returns available seasons for a given league in the Opta data.
}
\examples{
\dontrun{
list_opta_seasons("ENG")
list_opta_seasons("EPL")
}
}
