% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duel_model.R
\name{load_duel_model}
\alias{load_duel_model}
\title{Load the pre-trained xDuel model}
\usage{
load_duel_model(path = NULL)
}
\arguments{
\item{path}{Optional explicit path. Falls back to pannamodels, then local.}
}
\value{
Fitted xDuel model.
}
\description{
Load the pre-trained xDuel model
}
