% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{load_epv_model}
\alias{load_epv_model}
\title{Load EPV Model}
\usage{
load_epv_model(path = NULL)
}
\arguments{
\item{path}{Directory containing model. If NULL, uses pannadata/models/opta/}
}
\value{
EPV model
}
\description{
Loads pre-trained EPV model from disk.
}
\examples{
\dontrun{
epv_model <- load_epv_model()
epv_model <- load_epv_model(path = "models/custom_dir")
}

}
