% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_model.R
\name{load_wp_model}
\alias{load_wp_model}
\title{Load WP model}
\usage{
load_wp_model(path = NULL)
}
\arguments{
\item{path}{Directory to load from. If NULL, uses \code{pannadata/data/opta/models/}.}
}
\value{
WP model list (model + feature_names).
}
\description{
Load WP model
}
