% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{load_xpass_model}
\alias{load_xpass_model}
\title{Load Pre-trained xPass Model}
\usage{
load_xpass_model(path = NULL)
}
\arguments{
\item{path}{Path to model RDS file. If NULL, uses default location.}
}
\value{
Fitted xPass model
}
\description{
Loads xPass model from saved RDS file.
}
