% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{match_is_international}
\alias{match_is_international}
\title{Classify competitions as international vs domestic}
\usage{
match_is_international(league)
}
\arguments{
\item{league}{Character vector of competition codes.}
}
\value{
Logical vector -- \code{TRUE} for international (national-team)
competitions, \code{FALSE} for domestic club competitions.
}
\description{
Classify competitions as international vs domestic
}
\seealso{
Other world cup simulation: 
\code{\link{bt_match_prob}()},
\code{\link{build_knockout_lookup}()},
\code{\link{compute_league_offsets}()},
\code{\link{fit_bt_ratings}()},
\code{\link{mirror_match_rows}()},
\code{\link{run_wc2026_reference_checks}()},
\code{\link{simulate_world_cup}()}
}
\concept{world cup simulation}
