% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_GAMES_FOR_PADDING}
\alias{MIN_GAMES_FOR_PADDING}
\title{Default minimum games for Bayesian padding full weight}
\format{
Integer value: 10
}
\usage{
MIN_GAMES_FOR_PADDING
}
\description{
Number of games required for a player's statistics to receive full weight
in Bayesian padding calculations. Players with fewer games are regressed
toward population mean.
}
\examples{
MIN_GAMES_FOR_PADDING
}
\keyword{datasets}
