% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_MINUTES_FEATURES}
\alias{MIN_MINUTES_FEATURES}
\title{Default minimum minutes for player feature matrix}
\format{
Integer value: 180
}
\usage{
MIN_MINUTES_FEATURES
}
\description{
Minimum minutes for inclusion in player feature matrices used
for SPM prediction and Panna rating calculation.
}
\examples{
MIN_MINUTES_FEATURES
}
\keyword{datasets}
