% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_MINUTES_RAPM}
\alias{MIN_MINUTES_RAPM}
\title{Default minimum minutes for RAPM matrix inclusion}
\format{
Integer value: 90
}
\usage{
MIN_MINUTES_RAPM
}
\description{
Players must have at least this many total minutes to be included
in the RAPM design matrix as individual players. Players below this
threshold are grouped into a "replacement" pool.
}
\examples{
MIN_MINUTES_RAPM
}
\keyword{datasets}
