% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_MINUTES_SPM}
\alias{MIN_MINUTES_SPM}
\title{Default minimum minutes for SPM/player stats functions}
\format{
Integer value: 450
}
\usage{
MIN_MINUTES_SPM
}
\description{
Minimum minutes threshold used by player stats aggregation functions
and SPM model training. Higher than RAPM threshold because box score
stats need more sample size for stability.
}
\examples{
MIN_MINUTES_SPM
}
\keyword{datasets}
