% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_SEQUENCES_PER_MATCH}
\alias{MIN_SEQUENCES_PER_MATCH}
\title{Minimum estimated sequences per team per match}
\format{
Integer value: 20
}
\usage{
MIN_SEQUENCES_PER_MATCH
}
\description{
Lower bound for sequence estimation. A team will have at least
this many sequences per match regardless of touch count.
}
\examples{
MIN_SEQUENCES_PER_MATCH
}
\keyword{datasets}
