% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_SHOTS_FOR_FINISHING}
\alias{MIN_SHOTS_FOR_FINISHING}
\title{Default minimum shots for finishing modifier calculation}
\format{
Integer value: 20
}
\usage{
MIN_SHOTS_FOR_FINISHING
}
\description{
Minimum shots required for a player to have a meaningful finishing
modifier calculated. Players with fewer shots get default modifier.
}
\examples{
MIN_SHOTS_FOR_FINISHING
}
\seealso{
Other constants: 
\code{\link{BETA_PRIOR_ALPHA}},
\code{\link{CHAIN_TIME_GAP_SECONDS}},
\code{\link{CONFIDENCE_LEVEL}},
\code{\link{HALFTIME_MINUTE}},
\code{\link{MINUTES_PER_MATCH}},
\code{\link{MIN_90S_PSR_LEADERBOARD}},
\code{\link{MIN_GAMES_FOR_PADDING}},
\code{\link{MIN_MINUTES_FEATURES}},
\code{\link{MIN_MINUTES_RAPM}},
\code{\link{MIN_MINUTES_SPM}},
\code{\link{MIN_SEQUENCES_PER_MATCH}},
\code{\link{MIN_WEIGHT_DURATION}},
\code{\link{PENALTY_SHOOTOUT_CONVERSION}},
\code{\link{PENALTY_XG}},
\code{\link{PLAYERS_PER_TEAM}},
\code{\link{SIX_YARD_X_MIN}},
\code{\link{SIX_YARD_Y_MAX}},
\code{\link{SIX_YARD_Y_MIN}},
\code{\link{TOUCHES_PER_SEQUENCE}},
\code{\link{XG_MAX}},
\code{\link{XG_MIN}}
}
\concept{constants}
\keyword{datasets}
