% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_SHOTS_FOR_FINISHING}
\alias{MIN_SHOTS_FOR_FINISHING}
\title{Default minimum shots for finishing modifier calculation}
\format{
Integer value: 20
}
\usage{
MIN_SHOTS_FOR_FINISHING
}
\description{
Minimum shots required for a player to have a meaningful finishing
modifier calculated. Players with fewer shots get default modifier.
}
\examples{
MIN_SHOTS_FOR_FINISHING
}
\keyword{datasets}
