% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_WEIGHT_DURATION}
\alias{MIN_WEIGHT_DURATION}
\title{Minimum weight threshold for duration-based weighting}
\format{
Numeric value: 0.01
}
\usage{
MIN_WEIGHT_DURATION
}
\description{
Floor value for weights to prevent division by very small numbers
in RAPM weighting. Splints are weighted by minutes/90, with this
as the minimum.
}
\examples{
MIN_WEIGHT_DURATION
}
\keyword{datasets}
