% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MINUTES_PER_MATCH}
\alias{MINUTES_PER_MATCH}
\title{Minutes per regulation match}
\format{
Integer value: 90
}
\usage{
MINUTES_PER_MATCH
}
\description{
Standard duration of a regulation football match (90 minutes).
Used for per-90 rate calculations.
}
\examples{
MINUTES_PER_MATCH
}
\keyword{datasets}
