% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna-package.R
\docType{package}
\name{panna-package}
\alias{panna}
\alias{panna-package}
\title{panna: Player Rating System for Football Using RAPM and SPM Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements player ratings for football (soccer) using Regularized Adjusted Plus-Minus ('RAPM') combined with Statistical Plus-Minus ('SPM') as a Bayesian prior. Analyses time segments with constant lineups ('splints') to estimate each player's contribution to expected goals ('xG') differential. Provides offensive and defensive rating decomposition across multiple data sources including 'FBref', 'Opta', and 'Understat'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/peteowen1/panna}
  \item Report bugs at \url{https://github.com/peteowen1/panna/issues}
}

}
\author{
\strong{Maintainer}: Pete Owen \email{pete.owen1@hotmail.co.uk} [copyright holder]

}
\keyword{internal}
