% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna-package.R
\docType{package}
\name{panna-package}
\alias{panna}
\alias{panna-package}
\title{panna: Player Rating System for Football Using RAPM and SPM Methods}
\description{
Implements the panna player rating methodology for football (soccer) players. Combines Regularized Adjusted Plus-Minus (RAPM) with Statistical Plus-Minus (SPM) as a Bayesian prior to produce stable player impact ratings. Includes data collection from FBref via worldfootballR, splint-based analysis (time segments between substitutions/goals), and offensive/defensive rating separation.
}
\author{
\strong{Maintainer}: Pete Owen \email{pete.owen1@hotmail.co.uk}

}
\keyword{internal}
