% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PANNA_PSR_WEIGHT}
\alias{PANNA_PSR_WEIGHT}
\title{PSR weight in combined Panna Value rating}
\format{
Numeric value: 0.5
}
\usage{
PANNA_PSR_WEIGHT
}
\description{
PSR weight in combined Panna Value rating
}
\keyword{datasets}
