% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_conversion.R
\name{parse_opta_qualifiers}
\alias{parse_opta_qualifiers}
\title{Parse Opta Qualifiers JSON}
\usage{
parse_opta_qualifiers(events)
}
\arguments{
\item{events}{Data frame with qualifier_json column}
}
\value{
Data frame with parsed qualifier columns added
}
\description{
Extracts qualifier information from the qualifier_json column.
Opta JSON format is a dictionary where keys are qualifier IDs:
\code{{"108":null,"55":"145","28":null,...}}
}
\keyword{internal}
