% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{pb_download_epv_models}
\alias{pb_download_epv_models}
\title{Download EPV Models from GitHub Releases}
\usage{
pb_download_epv_models(
  repo = "peteowen1/pannadata",
  tag = "epv-models",
  dest = NULL
)
}
\arguments{
\item{repo}{GitHub repository (default: peteowen1/pannadata)}

\item{tag}{Release tag (default: epv-models)}

\item{dest}{Destination directory. If NULL, uses pannadata/data/opta/models/}
}
\value{
Invisibly returns path to models
}
\description{
Downloads pre-trained EPV models from GitHub releases.
}
\seealso{
Other epv: 
\code{\link{aggregate_player_xmetrics}()},
\code{\link{assign_epv_credit}()},
\code{\link{convert_opta_to_spadl}()},
\code{\link{enrich_match_stats_with_xmetrics}()},
\code{\link{fit_epv_model}()},
\code{\link{fit_xg_model}()},
\code{\link{fit_xgot_model}()},
\code{\link{fit_xpass_model}()},
\code{\link{get_or_build_spadl}()},
\code{\link{load_epv_model}()},
\code{\link{load_xg_model}()},
\code{\link{load_xgot_model}()},
\code{\link{load_xpass_model}()},
\code{\link{predict_xg}()},
\code{\link{predict_xgot}()},
\code{\link{predict_xpass}()},
\code{\link{save_epv_model}()},
\code{\link{save_xg_model}()},
\code{\link{save_xpass_model}()}
}
\concept{epv}
