% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{pb_download_epv_models}
\alias{pb_download_epv_models}
\title{Download EPV Models from GitHub Releases}
\usage{
pb_download_epv_models(
  repo = "peteowen1/pannadata",
  tag = "epv-models",
  dest = NULL
)
}
\arguments{
\item{repo}{GitHub repository (default: peteowen1/pannadata)}

\item{tag}{Release tag (default: epv-models)}

\item{dest}{Destination directory. If NULL, uses pannadata/models/opta/}
}
\value{
Invisibly returns path to models
}
\description{
Downloads pre-trained EPV models from GitHub releases.
}
\examples{
\dontrun{
pb_download_epv_models()
pb_download_epv_models(tag = "epv-models-v2")
}

}
