% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{pb_list_data}
\alias{pb_list_data}
\title{List files in GitHub Release}
\usage{
pb_list_data(repo = "peteowen/pannadata", tag = "data")
}
\arguments{
\item{repo}{GitHub repository in "owner/repo" format}

\item{tag}{Release tag (default: "data")}
}
\value{
Data frame with file information
}
\description{
Shows what data files are available in a GitHub Release.
}
\examples{
\dontrun{
pb_list_data()
}
}
