% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggyback.R
\name{pb_sync_data}
\alias{pb_sync_data}
\title{Sync local data with GitHub Releases}
\usage{
pb_sync_data(repo = "peteowen1/pannadata", tag = "latest", source = NULL)
}
\arguments{
\item{repo}{GitHub repository in "owner/repo" format}

\item{tag}{Release tag (default: "latest")}

\item{source}{Source directory (default: pannadata_dir())}
}
\value{
Invisible NULL
}
\description{
Convenience function that uploads local data to GitHub Releases.
This replaces the existing data in the release.
}
\examples{
\dontrun{
pb_sync_data()
}
}
