% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PENALTY_XG}
\alias{PENALTY_XG}
\title{Default penalty kick xG value}
\format{
Numeric value: 0.76
}
\usage{
PENALTY_XG
}
\description{
xG override for penalty kicks. Based on historical penalty
conversion rates (~76\%). Used when xG model is trained without
penalties.
}
\examples{
PENALTY_XG
}
\keyword{datasets}
