% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{per_90}
\alias{per_90}
\title{Calculate minutes per 90}
\usage{
per_90(stat, minutes)
}
\arguments{
\item{stat}{Statistic value (counting stat)}

\item{minutes}{Minutes played}
}
\value{
Statistic per 90 minutes
}
\description{
Calculate minutes per 90
}
\examples{
per_90(5, 450)
per_90(c(3, 7), c(270, 810))
}
