% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_world_cup.R
\name{play_knockout_round}
\alias{play_knockout_round}
\title{Play one round of a knockout bracket (internal helper)}
\usage{
play_knockout_round(bracket, r_vec, team_idx, nu)
}
\description{
Play one round of a knockout bracket (internal helper)
}
\keyword{internal}
