% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PLAYERS_PER_TEAM}
\alias{PLAYERS_PER_TEAM}
\title{Players per team in standard lineup}
\format{
Integer value: 11
}
\usage{
PLAYERS_PER_TEAM
}
\description{
Number of players per team in a standard football lineup.
}
\examples{
PLAYERS_PER_TEAM
}
\keyword{datasets}
