% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{predict_xpass}
\alias{predict_xpass}
\title{Predict Pass Completion Probability}
\usage{
predict_xpass(xpass_model, pass_features)
}
\arguments{
\item{xpass_model}{Fitted xPass model from fit_xpass_model()}

\item{pass_features}{Data frame with pass features}
}
\value{
Vector of xPass predictions (probabilities)
}
\description{
Applies trained xPass model to predict completion probability for passes.
}
\examples{
\dontrun{
xpass_model <- load_xpass_model()
pass_features <- prepare_passes_for_xpass(spadl_actions)
xpass_pred <- predict_xpass(xpass_model, pass_features)
}
}
\seealso{
Other epv: 
\code{\link{aggregate_player_xmetrics}()},
\code{\link{assign_epv_credit}()},
\code{\link{convert_opta_to_spadl}()},
\code{\link{enrich_match_stats_with_xmetrics}()},
\code{\link{fit_epv_model}()},
\code{\link{fit_xg_model}()},
\code{\link{fit_xgot_model}()},
\code{\link{fit_xpass_model}()},
\code{\link{get_or_build_spadl}()},
\code{\link{load_epv_model}()},
\code{\link{load_xg_model}()},
\code{\link{load_xgot_model}()},
\code{\link{load_xpass_model}()},
\code{\link{pb_download_epv_models}()},
\code{\link{predict_xg}()},
\code{\link{predict_xgot}()},
\code{\link{save_epv_model}()},
\code{\link{save_xg_model}()},
\code{\link{save_xpass_model}()}
}
\concept{epv}
