% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{predict_xpass}
\alias{predict_xpass}
\title{Predict Pass Completion Probability}
\usage{
predict_xpass(xpass_model, pass_features)
}
\arguments{
\item{xpass_model}{Fitted xPass model from fit_xpass_model()}

\item{pass_features}{Data frame with pass features}
}
\value{
Vector of xPass predictions (probabilities)
}
\description{
Applies trained xPass model to predict completion probability for passes.
}
