% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_all_data}
\alias{process_all_data}
\title{Process all raw data}
\usage{
process_all_data(raw_data)
}
\arguments{
\item{raw_data}{List of raw data from scrape_pl_comprehensive}
}
\value{
List of processed data frames
}
\description{
Master function to process all collected data.
}
