% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{rank_players_panna}
\alias{rank_players_panna}
\title{Rank players by panna rating}
\usage{
rank_players_panna(ratings, top_n = NULL, position = NULL)
}
\arguments{
\item{ratings}{Data frame of panna ratings}

\item{top_n}{Number of top players to show}

\item{position}{Optional position filter}
}
\value{
Data frame of ranked players
}
\description{
Returns ranked list of players.
}
\examples{
\dontrun{
ratings <- get_panna_ratings(panna_model)
top_20 <- rank_players_panna(ratings, top_n = 20)
forwards <- rank_players_panna(ratings, top_n = 10, position = "FW")
}

}
