% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe division handling division by zero}
\usage{
safe_divide(x, y, default = 0)
}
\arguments{
\item{x}{Numerator}

\item{y}{Denominator}

\item{default}{Value to return when denominator is zero (default: 0)}
}
\value{
x / y, or default if y is zero
}
\description{
Safe division handling division by zero
}
\examples{
safe_divide(10, 2)
safe_divide(10, 0)
safe_divide(10, 0, default = NA)
}
