% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_model.R
\name{save_epv_model}
\alias{save_epv_model}
\title{Save EPV Model}
\usage{
save_epv_model(epv_model, path = NULL)
}
\arguments{
\item{epv_model}{EPV model from fit_epv_model()}

\item{path}{Directory to save model. If NULL, uses pannadata/opta/models/}
}
\value{
Invisibly returns path
}
\description{
Saves trained EPV model to disk.
}
