% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_model.R
\name{save_wp_model}
\alias{save_wp_model}
\title{Save WP model}
\usage{
save_wp_model(wp_model, path = NULL)
}
\arguments{
\item{wp_model}{WP model from \code{\link{train_wp_model}}.}

\item{path}{Directory to save. If NULL, uses \code{pannadata/data/opta/models/}.}
}
\value{
Invisibly returns the file path.
}
\description{
Save WP model
}
