% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_model.R
\name{save_xg_model}
\alias{save_xg_model}
\title{Save xG Model}
\usage{
save_xg_model(xg_model, path = NULL)
}
\arguments{
\item{xg_model}{Fitted xG model from fit_xg_model()}

\item{path}{Path to save model. If NULL, saves to pannadata/opta/models/}
}
\value{
Invisibly returns the path
}
\description{
Saves trained xG model to RDS file.
}
\seealso{
Other epv: 
\code{\link{aggregate_player_xmetrics}()},
\code{\link{assign_epv_credit}()},
\code{\link{convert_opta_to_spadl}()},
\code{\link{enrich_match_stats_with_xmetrics}()},
\code{\link{fit_epv_model}()},
\code{\link{fit_xg_model}()},
\code{\link{fit_xgot_model}()},
\code{\link{fit_xpass_model}()},
\code{\link{get_or_build_spadl}()},
\code{\link{load_epv_model}()},
\code{\link{load_xg_model}()},
\code{\link{load_xgot_model}()},
\code{\link{load_xpass_model}()},
\code{\link{pb_download_epv_models}()},
\code{\link{predict_xg}()},
\code{\link{predict_xgot}()},
\code{\link{predict_xpass}()},
\code{\link{save_epv_model}()},
\code{\link{save_xpass_model}()}
}
\concept{epv}
