% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass_model.R
\name{save_xpass_model}
\alias{save_xpass_model}
\title{Save xPass Model}
\usage{
save_xpass_model(xpass_model, path = NULL)
}
\arguments{
\item{xpass_model}{Fitted xPass model}

\item{path}{Path to save. If NULL, saves to pannadata/opta/models/}
}
\value{
Invisibly returns the path
}
\description{
Saves trained xPass model to RDS file.
}
\examples{
\dontrun{
xpass_model <- fit_xpass_model(pass_features)
save_xpass_model(xpass_model)
save_xpass_model(xpass_model, path = "models/my_xpass_model.rds")
}
}
\seealso{
Other epv: 
\code{\link{aggregate_player_xmetrics}()},
\code{\link{assign_epv_credit}()},
\code{\link{convert_opta_to_spadl}()},
\code{\link{enrich_match_stats_with_xmetrics}()},
\code{\link{fit_epv_model}()},
\code{\link{fit_xg_model}()},
\code{\link{fit_xgot_model}()},
\code{\link{fit_xpass_model}()},
\code{\link{get_or_build_spadl}()},
\code{\link{load_epv_model}()},
\code{\link{load_xg_model}()},
\code{\link{load_xgot_model}()},
\code{\link{load_xpass_model}()},
\code{\link{pb_download_epv_models}()},
\code{\link{predict_xg}()},
\code{\link{predict_xgot}()},
\code{\link{predict_xpass}()},
\code{\link{save_epv_model}()},
\code{\link{save_xg_model}()}
}
\concept{epv}
