% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIX_YARD_X_MIN}
\alias{SIX_YARD_X_MIN}
\title{Six-yard box x threshold (attacking end)}
\format{
Numeric value: 94
}
\usage{
SIX_YARD_X_MIN
}
\description{
Minimum x coordinate for the attacking six-yard box on a 0-100 pitch.
}
\examples{
SIX_YARD_X_MIN
}
\keyword{datasets}
