% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIX_YARD_Y_MAX}
\alias{SIX_YARD_Y_MAX}
\title{Six-yard box y upper bound}
\format{
Numeric value: 63
}
\usage{
SIX_YARD_Y_MAX
}
\description{
Upper y boundary of the six-yard box on a 0-100 pitch.
}
\examples{
SIX_YARD_Y_MAX
}
\keyword{datasets}
