% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIX_YARD_Y_MIN}
\alias{SIX_YARD_Y_MIN}
\title{Six-yard box y lower bound}
\format{
Numeric value: 37
}
\usage{
SIX_YARD_Y_MIN
}
\description{
Lower y boundary of the six-yard box on a 0-100 pitch.
}
\examples{
SIX_YARD_Y_MIN
}
\keyword{datasets}
