% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{split_od_contributions}
\alias{split_od_contributions}
\title{Split O/D contributions from existing ratings}
\usage{
split_od_contributions(panna_ratings, player_features)
}
\arguments{
\item{panna_ratings}{Overall panna ratings}

\item{player_features}{Player features with offensive/defensive stats}
}
\value{
Data frame with estimated O-panna and D-panna
}
\description{
Approximates O/D split from overall panna and feature data.
}
\examples{
\dontrun{
features <- create_player_feature_matrix(processed_data)
od_ratings <- split_od_contributions(panna_ratings, features)
head(od_ratings[, c("player_name", "panna", "o_panna", "d_panna")])
}
}
