% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOUCHES_PER_SEQUENCE}
\alias{TOUCHES_PER_SEQUENCE}
\title{Average touches per possession sequence (approximation)}
\format{
Integer value: 5
}
\usage{
TOUCHES_PER_SEQUENCE
}
\description{
Rough approximation used to estimate the number of possession sequences
from total touches. Used in per-100-sequences rate calculations.
}
\examples{
TOUCHES_PER_SEQUENCE
}
\keyword{datasets}
