% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panna_rating.R
\name{validate_panna_ratings}
\alias{validate_panna_ratings}
\title{Validate panna ratings}
\usage{
validate_panna_ratings(panna_ratings, rapm_ratings, spm_ratings)
}
\arguments{
\item{panna_ratings}{Panna ratings}

\item{rapm_ratings}{RAPM ratings}

\item{spm_ratings}{SPM ratings}
}
\value{
List of validation metrics
}
\description{
Calculates validation metrics for panna ratings.
}
\examples{
\dontrun{
validation <- validate_panna_ratings(panna_ratings, rapm_ratings, spm_ratings)
validation$panna_rapm_cor
validation$panna_spm_cor
}

}
