% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_defensive.R
\name{visualize_od_scatter}
\alias{visualize_od_scatter}
\title{Visualize O/D scatter}
\usage{
visualize_od_scatter(ratings, highlight_top = 10)
}
\arguments{
\item{ratings}{O/D ratings data frame}

\item{highlight_top}{Number of top players to label}
}
\value{
ggplot object
}
\description{
Creates scatter plot of offensive vs defensive ratings.
Requires ggplot2.
}
\examples{
\dontrun{
od_result <- calculate_od_panna(rapm_data, spm_ratings)
visualize_od_scatter(od_result$ratings, highlight_top = 15)
}
}
