% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{XG_MAX}
\alias{XG_MAX}
\title{Maximum xG value (caps extreme predictions)}
\format{
Numeric value: 0.75
}
\usage{
XG_MAX
}
\description{
Ceiling for xG predictions to prevent overconfident predictions.
Even penalty kicks and open-net chances rarely exceed this threshold.
}
\examples{
XG_MAX
}
\keyword{datasets}
