% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{XG_MIN}
\alias{XG_MIN}
\title{Minimum xG value (prevents 0 in log calculations)}
\format{
Numeric value: 0.01
}
\usage{
XG_MIN
}
\description{
Floor for xG predictions to prevent issues with log calculations
and overly confident predictions of 0 probability.
}
\examples{
XG_MIN
}
\keyword{datasets}
